refactor(xpu): split Intel MM weights into module
